requirements.txt
flask
flask-wtf
flask-security
Flask-SQLAlchemy
WTForms
psycopg2-binary
python-dotenv
passlib
dataclasses
msal